requirements.txt

tensorflow==2.12.0
uvicorn==0.21.1
fastapi==0.95.2
keras==2.12.0
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.5.0
pydantic==1.10.7
Requests==2.30.0
